Фикс типа чисел и пробелов

+++ Lesson1/Triangle_area.rb
@@ -5,17 +5,16 @@ p "Hello, my name is #{programm_name}. I help you to find area of tringle."
 p "Need only 2 parameters for this. So tell me BASE of triangle in santimetres"
 
 #take user's input and save to 'base' variable
-base = gets.chomp.to_i
+base = gets.chomp.to_f
 
 p "Fine. Tell me the second parameter - HEIGHT"
 
 #takes user's input and save to 'height' variable
-height = gets.chomp.to_i
+height = gets.chomp.to_f
 
 # calculate area and save to variable
-area = 0.5*base*height
+area = 0.5 * base * height
 
 p "Ok, wait a little, please"
 sleep(3)
 p "Well done! Area of triangle is equal to #{area} !"
-
